pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Production-ready RAG-powered medical question answering chatbot built with LLMs, LangChain, Pinecone, Groq, and Flask."
readme = "README.md"
requires-python = ">=3.12,<3.13"

authors = [
  { name = "Omar Nagy", email = "onagy489@gmail.com" }
]

dependencies = [
  "langchain>=0.2.0,<0.3.0",
  "langchain-community>=0.2.0,<0.3.0",
  "langchain-openai>=0.2.0,<0.3.0",
  "langchain-pinecone>=0.1.0,<0.2.0",
  "langchain-groq>=0.1.0,<0.2.0",
  "flask>=3.0.0,<4.0.0",
  "sentence-transformers>=3.0.0,<4.0.0",
  "pypdf>=4.0.0,<5.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "ipynb>=0.5.0,<0.6.0",
  "langchain_huggingface>=0.0.3,<0.1.0",
]


